+++ extension/models/Errors/FieldValidationError.js
@@ -5,24 +5,10 @@ const EFIELDVALIDATION = 'EVALIDATION'
  */
 class FieldValidationError extends Error {
   constructor (message) {
-    super('Field validation error')
+    super()
     this.code = EFIELDVALIDATION
     this.message = message || 'There was an error with the request'
     this.validationErrors = []
-  }
-
-  /**
-   * @param {string} path - path of the error of the field, e.g. firstName
-   * @param {string} message - error message that pertains to path
-   * @param {string} [value] - passed down value if any
-   */
-  addValidationMessage (path, message, value) {
-    path = this.translatePath(path)
-
-    const capitalizedPath = path.charAt(0).toUpperCase() + path.substr(1)
-    const main = value && !message.includes('required') ? value : capitalizedPath
-
-    this.validationErrors.push({ path, message: `"${main}" ${message}` })
   }
 
   /**