+++ extension/user/address/update.js
@@ -1,4 +1,3 @@
-const Tools = require('../../lib/tools')
 const UnauthorizedError = require('../../models/Errors/UnauthorizedError')
 const FieldValidationError = require('../../models/Errors/FieldValidationError')
 const AddressValidationError = require('../../models/Errors/AddressValidationError')
@@ -7,11 +6,14 @@ const _ = require('lodash')
 const ApiFactory = require('../../lib/shopify.api.factory')
 
 /**
+ * @typedef {Object} input
+ * @property {string} id - id of Shopify address to update
+ *
  * @param {SDKContext} context
- * @param {ShopgateAddress} input
+ * @param input
  */
 module.exports = async function (context, input) {
-  if (Tools.isEmpty(context.meta.userId)) {
+  if (!context.meta.userId) {
     throw new UnauthorizedError('User is not logged in.')
   }
 
@@ -19,31 +21,30 @@ module.exports = async function (context, input) {
   const storefrontApi = ApiFactory.buildStorefrontApi(context, storeFrontAccessToken)
   const customerAccessToken = await context.storage.user.get('customerAccessToken')
 
-  return storefrontApi.customerAddressUpdate(customerAccessToken.accessToken, input.id, createAddressUpdate(input)).then(result => {
-    return { success: true }
-  }).catch(errors => {
-    const validationError = new FieldValidationError()
-    if (Array.isArray(errors)) {
-      errors.forEach(error => {
-        const { message } = error
-        if (Tools.propertyExists(error, 'problems')) {
-          throw new AddressValidationError(message)
-        }
-        const { field } = error
-        if (field[1]) {
-          validationError.addStorefrontValidationMessage(field[1], message)
-        }
-      })
-    }
-    if (validationError.validationErrors.length > 0) {
-      throw validationError
-    }
-  })
+  return storefrontApi.customerAddressUpdate(customerAccessToken.accessToken, input.id, createAddressUpdate(input))
+    .catch(errors => {
+      const validationError = new FieldValidationError()
+      if (Array.isArray(errors)) {
+        errors.forEach(error => {
+          const { message } = error
+          if (error.hasOwnProperty('problems')) {
+            throw new AddressValidationError(message)
+          }
+          const { field } = error
+          if (field[1]) {
+            validationError.addStorefrontValidationMessage(field[1], message)
+          }
+        })
+      }
+      if (validationError.validationErrors.length > 0) {
+        throw validationError
+      }
+    })
 
   /**
-   * Map the input address values to fit the Shopify specifications for the admin api endpoint
+   * Map the input address values to fit the Shopify specifications for the api endpoint
    * @param {ShopgateAddress} input
-   * @returns {Object}
+   * @returns {ShopifyAddress}
    */
   function createAddressUpdate (input) {
     const newAddress = {