+++ extension/user/token/storeWebLoginPhrase.js
@@ -1,26 +1,10 @@
 /**
- * @typedef {object} input
+ * @typedef {Object} input
  * @property {string} phrase
  *
- * @param context
+ * @param {SDKContext} context
  * @param input
- * @param cb
- * @returns {*}
  */
-module.exports = function (context, input, cb) {
-  saveWebLoginPhrase(input.phrase, context, cb)
-}
-
-/**
- * Saves the given web login phrase into internal device storage
- *
- * @param {string} phrase
- * @param {Object} context
- * @param {function({Object})} cb
- */
-function saveWebLoginPhrase (phrase, context, cb) {
-  // use device storage only
-  context.storage.device.set('webLoginPhrase', phrase, (err) => {
-    return cb(err || null)
-  })
+module.exports = async (context, input) => {
+  context.storage.device.set('webLoginPhrase', input.phrase)
 }